changing 'no new codes' tests to integration tests
