+++ tests/api_tests/test_questionnaire_api.py
@@ -1,7 +1,7 @@
 import http.client
 import json
 
-from rdr_service.code_constants import PPI_EXTRA_SYSTEM
+from rdr_service.code_constants import PPI_EXTRA_SYSTEM, PPI_SYSTEM
 from rdr_service.dao.code_dao import CodeDao
 from tests.test_data import data_path
 from tests.helpers.unittest_base import BaseTestCase
@@ -112,4 +112,43 @@ class QuestionnaireApiTest(BaseTestCase):
 
         # Ensure we didn't create codes in the extra system
         self.assertIsNone(CodeDao().get_code(PPI_EXTRA_SYSTEM, 'IgnoreThis'))
+
+    def test_questionnaire_payload_cannot_create_new_codes(self):
+        response = self.send_post('Questionnaire', {
+            "group": {
+                "concept": [
+                    {
+                        "system": PPI_SYSTEM,
+                        "code": 'new_module_code'
+                    }
+                ],
+                "question": [
+                    {
+                        "linkId": "test",
+                        "concept": [
+                            {
+                                "code": "new_question_code",
+                                "system": PPI_SYSTEM
+                            }
+                        ],
+                        "option": [
+                            {
+                                "code": "new_answer_code",
+                                "system": PPI_SYSTEM
+                            }
+                        ]
+                    }
+                ]
+            },
+            "status": "published",
+            "version": "V1"
+        }, expected_status=400)
+
+        self.assertEqual(
+            "The following code values were unrecognized: "
+            "new_module_code (system: http://terminology.pmi-ops.org/CodeSystem/ppi), "
+            "new_question_code (system: http://terminology.pmi-ops.org/CodeSystem/ppi), "
+            "new_answer_code (system: http://terminology.pmi-ops.org/CodeSystem/ppi)",
+            response.json['message']
+        )
